requirements.txt
black
cryptography==2.7
Django==2.2.4
djangorestframework==3.10.2
psycopg2==2.8.3
django-cors-headers
gunicorn 
whitenoise
